LINK
Click on Files Changed.

Click on the ellipsis (...) and select Edit file.

In the list you just created, add a link to each of your favorite sites to their respective URLs. You can do this by adding an opening anchor tag <a> tag with href attribute with your favorite site's URL, the name of the site inside the anchor tag, and a closing anchor </a> tag. Here is an example of a list item with a link: